+++ packages/demo/.env
@@ -1,1 +1,1 @@
-BROWSE=none
+BROWSER=none
